requirements.txt
loguru==0.7.2
python_dateutil==2.9.0.post0
setuptools==69.5.1
